refactor(etl): выделить ETL-сервис в отдельный модуль
• Переместить ETL-сервис из `bot/etl_service/` в корневой `etl_service/`
• Удалить настройки аналитической БД из `bot/settings.py` и перенести в `etl_service/config.py`
• Добавить Dockerfile и requirements.txt для автономного развертывания ETL-сервиса
• Обновить пути импортов и конфигурацию внутри ETL-модуля
