+++ bot/settings.py
@@ -34,13 +34,6 @@ class Settings(BaseSettings):
     # Logging configuration
     log_min_level: str = Field("INFO", description="Minimum logging level (DEBUG, INFO, WARNING, ERROR)")
 
-    # Anal DB
-    anal_postgres_db: str | None = None
-    anal_postgres_user: str | None = None
-    anal_postgres_password: str | None = None
-    anal_postgres_port: str | None = None
-    anal_postgres_host: str | None = None
-
     class Config:
         env_file = ".env"
         case_sensitive = True
