+++ etl_service/requirements.txt
@@ -1,0 +1,3 @@
+apscheduler>=3.11.2
+pydantic-settings==2.12.0
+asyncpg==0.31.0
